Restructured project and rewrite mockups to json format
