+++ src/sequelize/mockups/accessorie.json
@@ -3,7 +3,7 @@
     "travelBoxes": true,
     "lock": true,
     "camera": true,
-    "basicTools": "included",
+    "basicTools": true,
     "firstAidKit": true
   }
 ]
