+++ src/sequelize/mockups/index.ts
@@ -1,80 +1,62 @@
-import {
-  Location,
-  Offer,
-  Offerer,
-  Tag,
-  User,
-} from '../';
-import sequelize from '../config';
-import locations from './locations';
-import offers from './offers';
-import offerers from './offerers';
-import tags from './tags';
-import users from './users';
+import * as walkSync from 'walk-sync';
+import * as appRoot from 'app-root-path';
+import * as fs from 'fs';
 import debug from 'debug';
-const log = debug('api:Sequelize');
+import db from '../../motocycle';
+const log = debug('api:sequelize');
 
-export function createLocations(): Promise<any> {
-  return Location.sync({ force: true })
-    .then(() => Location.bulkCreate(locations))
-    .catch(err => log('Database bulkCreate error', err));
-}
+export default function fillDb() {
+  const path = `${appRoot.path}/src/sequelize/mockups/motorcycle`
+  const paths = walkSync(`${path}`, { ignore: ['index.ts'] });
+  const promises = [];
 
-export function createUsers(): Promise<any> {
-  return User.sync({ force: true })
-    .then(() => User.bulkCreate(users))
-    .catch(err => log('Database bulkCreate error', err));
-}
+  paths.forEach((file) => {
+    const json = fs.readFileSync(`${path}/${file}`, 'utf-8');
+    const data = JSON.parse(json);
+    const model = file.replace('.json', '');
+    promises.push(db[model].bulkCreate(data).catch(err => console.log('Bulk create', err)));
+  })
 
-export function createOffers(): Promise<any> {
-  return Offer.sync({ force: true })
-    .then(() => Offer.bulkCreate(offers))
-    .catch(err => log('Database bulkCreate error', err));
-}
+  Promise.all(promises).then(() => {
+    const motorcycle = db['motorcycle']
+    motorcycle.findById(1)
+      .then((moto) => {
+        Promise.all([
+          moto.setAccessorie(1),
+          moto.setMake('KTM'),
+          moto.setModel('SX 125'),
+          moto.setModelspec(1),
+          moto.setMotorspec(1),
+          moto.setProtection(1),
+          moto.setService(1),
+        ]).then(() => log(moto.dataValues));
+      })
+      .catch(err => console.log('All error', err));
 
-export function createOfferers(): Promise<any> {
-  return Offerer.sync({ force: true })
-    .then(() => Offerer.bulkCreate(offerers))
-    .catch(err => log('Database bulkCreate error', err));
-}
+    motorcycle.findById(2)
+      .then((moto) => {
+        Promise.all([
+          moto.setAccessorie(1),
+          moto.setMake('Husaberg'),
+          moto.setModel('FE 390'),
+          moto.setModelspec(1),
+          moto.setMotorspec(1),
+          moto.setProtection(1),
+          moto.setService(1),
+        ]).then(() => log(moto.dataValues));
+      });
 
-export function createTags(): Promise<any> {
-  return Tag.sync({ force: true })
-    .then(() => Tag.bulkCreate(tags))
-    .catch(err => log('Database bulkCreate error', err));
-}
-
-export function createRelations(): Promise<any> {
-
-  Offer.belongsToMany(Tag, { through: 'OfferTag' });
-  Tag.belongsToMany(Offer, { through: 'OfferTag' });
-
-  Offer.belongsTo(Location);
-  Offer.belongsTo(Offerer);
-
-  return sequelize.sync({ force: true });
-}
-
-export function createAll() {
-  createRelations().then(() => {
-    Promise.all([
-      createLocations(),
-      createUsers(),
-      createOffers(),
-      createOfferers(),
-      createTags(),
-    ]).then(() => {
-        Offer.findById(1)
-          .then((offer) => {
-            Promise.all([
-              offer.setTags([3, 5, 7, 8]),
-              offer.setLocation(1),
-              offer.setOfferer(1),
-            ]).then(() => {
-              console.log(offer);
-            });
-          });
-        });
+    motorcycle.findById(3)
+      .then((moto) => {
+        Promise.all([
+          moto.setAccessorie(1),
+          moto.setMake('KTM'),
+          moto.setModel('Freeride 250R'),
+          moto.setModelspec(1),
+          moto.setMotorspec(1),
+          moto.setProtection(1),
+          moto.setService(1),
+        ]).then(() => log(moto.dataValues));
+      });
   });
 }
-
