+++ src/sequelize/mockups/locations.json
@@ -1,0 +1,11 @@
+[
+  {
+    "district": "mazowieckie",
+    "city": "Warszawa",
+    "area": "Śródmieście",
+    "country": "Poland",
+    "latitude": 52.237671,
+    "longitude": 21.030145,
+    "placeId": "ChIJaxh_0VnMHkcRKm5GlN_e0mM"
+  }
+]
