--- src/sequelize/mockups/locations.ts
@@ -1,29 +1,0 @@
-export default [
-  {
-    district: 'mazowieckie',
-    city: 'Warszawa',
-    area: 'Śródmieście',
-    country: 'Poland',
-    latitude: 52.237671,
-    longitude: 21.030145,
-    placeId: 'ChIJaxh_0VnMHkcRKm5GlN_e0mM',
-  },
-  {
-    district: 'mazowieckie',
-    city: 'Radziejowice',
-    area: 'Żyrardów',
-    country: 'Poland',
-    latitude: 52.048057,
-    longitude: 20.56829,
-    placeId: 'ChIJeTTB7ktBGUcROxxB7vwAZUs',
-  },
-  {
-    district: 'mazowieckie',
-    city: 'Zielonka',
-    area: 'Zielonka',
-    country: 'Poland',
-    latitude: 52.312614,
-    longitude: 21.184979,
-    placeId: 'ChIJMSPzCCHQHkcR9fZn-ovCSfU',
-  },
-];
