+++ src/sequelize/mockups/modelspec.json
