+++ src/sequelize/mockups/motorcycle.json
