+++ src/sequelize/mockups/motorspec.json
