+++ src/sequelize/mockups/offerers.json
@@ -1,0 +1,10 @@
+[
+  {
+    "name": "Michał Mikołajczyk",
+    "picture": "https://scontent-waw1-1.cdninstagram.com/t51.2885-15/e35/15877344_344661349253204_5404343355553349632_n.jpg"
+  },
+  {
+    "name": "Jan Horubała",
+    "picture": "https://scontent-amt2-1.cdninstagram.com/t51.2885-15/e35/16463985_256259144802357_4986204357729976320_n.jpg"
+  }
+]
