+++ src/sequelize/mockups/offers.json
@@ -1,0 +1,14 @@
+[
+  {
+    "title": "KTM SX 125 2012",
+    "type": "motocross"
+  },
+  {
+    "title": "Husaberg FE 390 2010",
+    "type": "motocross"
+  },
+  {
+    "title": "KTM Freeride 250R 2017",
+    "type": "motocross"
+  }
+]
