--- src/sequelize/mockups/offers.ts
@@ -1,14 +1,0 @@
-export default [
-  {
-    title: 'KTM SX 125 2012',
-    type: 'motocross',
-  },
-  {
-    title: 'Husaberg FE 390 2010',
-    type: 'motocross',
-  },
-  {
-    title: 'KTM Freeride 250R 2017',
-    type: 'motocross',
-  },
-];
