+++ src/sequelize/mockups/protection.json
