--- src/sequelize/mockups/tags.ts
@@ -1,12 +1,0 @@
-const types = [
-  'club',
-  'coach',
-  'guide',
-  'mechanic',
-  'motocycle',
-  'photographer',
-  'transport',
-  'garage',
-].map(name => ({ name }));
-
-export default types;
