--- src/sequelize/mockups/users.ts
@@ -1,23 +1,0 @@
-export default [
-  {
-    name: 'Viking Garage',
-    email: 'viking.garage.app@gmail.com',
-    birthday: new Date(1980, 12, 12),
-    password: 'secret',
-    verified: true,
-  },
-  {
-    name: 'Jan Horubala',
-    email: 'janek@vikinggarage.com',
-    birthday: new Date(1980, 12, 12),
-    password: 'secret',
-    verified: false,
-  },
-  {
-    name: 'Michal Mikolajczyk',
-    email: 'michal@vikinggarage.com',
-    birthday: new Date(1980, 12, 12),
-    password: 'secret',
-    verified: true,
-  },
-];
