+++ src/sequelize/models/model.ts
