+++ src/sequelize/models/modelspec.ts
