+++ src/sequelize/models/motorcycle.ts
