+++ src/sequelize/models/motorspec.ts
