+++ src/sequelize/models/protection.ts
